* 'feature-gh' of https://github.com/kosta-developer/DEVELOPER-Back: (22 commits)
  파일 다운로드 오류 해결(중복파일)
  세란님 꺼 병합하기전 작업부분 커밋푸시
  세란님 요청파일 추가
  이메일프로퍼티스삭제
  이메일프로퍼티스삭제
  [SR] 백단...
  병합전 커밋
  수정사항 반영
  세란님 수정요청 반영
  [SR] 메일API 적용
  세란님 요청으로 삭제
  세란님 요청으로 삭제
  Delete email.properties
  페이징처리 커밋+푸시
  Delete email.properties
  Delete email.properties
  Delete email.properties
  수정사항 반영
  removed cached
  장학님꺼 받기전 커밋 파일명 랜덤 변경
  ...

# Conflicts:
#	back/src/main/java/com/developer/email/EmailService.java
#	back/src/main/java/com/developer/email/EmailServiceImpl.java
#	back/src/main/java/com/developer/hostuser/repository/HostUserRepository.java
#	back/src/main/java/com/developer/hostuser/service/HostUserService.java
#	back/src/main/java/com/developer/join/JoinController.java
#	back/src/main/java/com/developer/lesson/dto/LessonDTO.java
#	back/src/main/java/com/developer/lesson/service/LessonService.java
#	back/src/main/java/com/developer/main/DeveloperController.java
#	back/src/main/java/com/developer/mypage/MyPageController.java
#	back/src/main/java/com/developer/users/dto/UsersDTO.java
#	back/src/main/java/com/developer/users/repository/UsersRepository.java
#	back/src/main/java/com/developer/util/FileDownloadController.java

+++ back/src/main/java/com/developer/hostuser/service/HostUserService.java
@@ -9,12 +9,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
+import com.developer.email.EmailService;
 import com.developer.exception.AddException;
 import com.developer.exception.FindException;
 import com.developer.exception.RemoveException;
 import com.developer.hostuser.dto.HostUserDTO;
 import com.developer.hostuser.entity.HostUser;
 import com.developer.hostuser.repository.HostUserRepository;
+
 
 import lombok.RequiredArgsConstructor;
 
@@ -23,6 +25,7 @@ import lombok.RequiredArgsConstructor;
 public class HostUserService {
 	
 	private final HostUserRepository hRepository;
+	private final EmailService emailService;
 	
 	ModelMapper modelMapper = new ModelMapper();
 	private Logger logger = LoggerFactory.getLogger(getClass());
@@ -168,16 +171,17 @@ public class HostUserService {
 	}
 
 	/**
-	 * 호스트 가입을 승인한다.
+	 * 호스트 가입을 승인한다. (메일 포함)
 	 * 
 	 * @author SR
 	 * @param hostId
-	 * @throws FindException
+	 * @throws FindException, Exception 
 	 */
-	public void readyOk(String hostId) throws FindException {
+	public void readyOk(String hostId) throws FindException, Exception {
 		Optional<HostUser> optHost = hRepository.findById(hostId);
 		HostUser hostEntity = optHost.get();
 		hostEntity.setReady(1);
+		emailService.hostOk(hostEntity.getEmail());
 		hRepository.save(hostEntity);
 	}
 
@@ -186,12 +190,18 @@ public class HostUserService {
 	 * 
 	 * @author SR
 	 * @param hostId
-	 * @throws RemoveException
+	 * @throws Exception 
+	 * @throws FindException 
 	 */
-	public void deleteHost(String hostId) throws RemoveException {
+	public void deleteHost(String hostId) throws FindException, Exception {
 		Optional<HostUser> optH = hRepository.findById(hostId);
-		HostUser entityH = optH.get();
-		hRepository.delete(entityH);
+		if(optH.isPresent()) {
+			HostUser entityH = optH.get();
+			emailService.hostReject(entityH.getEmail());
+			hRepository.delete(entityH);
+		} else {
+			throw new FindException("존재하지 않는 호스트 유저입니다.");
+		}
 	}
 	
 	/**
@@ -208,6 +218,7 @@ public class HostUserService {
 		if(optH.isPresent()) {
 			HostUser hostuser = optH.get();			
 			HostUserDTO.HostLoginDTO hostLoginDTO = modelMapper.map(hostuser, HostUserDTO.HostLoginDTO.class);
+			System.out.println("호스트아이디는 "+HostId+"비밀번호는"+pwd);
 			if(hostLoginDTO.getPwd().equals(pwd) && !hostLoginDTO.getReady().equals(2)) {
 				hostLoginDTO.setPwd("");
 				return hostLoginDTO;
@@ -250,5 +261,44 @@ public class HostUserService {
             return false; //가입된 정보가 있음
         }
    }
+	/**
+	 * 호스트 아이디 찾기
+	 * @author choigeunhyeong
+	 * @param Num 사업자 번호
+	 * @return
+	 * @throws FindException
+	 */
+	public HostUserDTO findHostId(String num) throws FindException{
+		Optional<HostUser> optH = hRepository.findByNum(num);
+		if(optH.isPresent()) {
+			HostUser host = optH.get();
+			HostUserDTO hostDTO = modelMapper.map(host, HostUserDTO.class);
+			if(hostDTO.getNum().equals(num)) {
+				return hostDTO;
+			}
+		}
+		throw new FindException("사업자번호가 일치하는 회원이 없습니다.");
+	}
 
+	/**
+	 * Email을 통해 해당 email로 가입된 정보가 있는지 확인. 
+	 * 가입된 정보가 있다면 입력받은 id와 email이 서로 일치한지 여부를 리턴하면서 임시비밀번호로 변경 및 메일발송
+	 * @param email
+	 * @param hostId
+	 * @return
+	 * @throws Exception
+	 */
+	public boolean hostPwdAndEmailCheck(String email, String hostId) throws Exception {
+		HostUser host = hRepository.hostEmailCheck(email);
+        if(host!=null && host.getHostId().equals(hostId)) {
+        	String temporaryPwd = emailService.updatePwd(email);
+        	host.setPwd(temporaryPwd);
+        	hRepository.save(host);
+        	return true;
+        }
+        else {
+            return false;
+        }
+	}
+	
 }
